+++ Servers/MapServer/EntityUpdateCodec.cpp
@@ -197,7 +197,7 @@ void sendSeqTriggeredMoves(const Entity &src,BitStream &bs)
         sCFDebug(logAnimations, "Sending seq triggered moves %" PRIu64, src.m_triggered_moves.size());
 
     // client appears to process only the last 20 triggered moves
-    bs.StorePackedBits(1, src.m_triggered_moves.size()); // num moves
+    bs.StorePackedBits(1, (uint32_t)src.m_triggered_moves.size()); // num moves
     for(const TriggeredMove &move : src.m_triggered_moves)
     {
         bs.StorePackedBits(10, move.m_move_idx);                   // 2  triggeredMoveIDX
