+++ Servers/MapServer/MessageHelpers.cpp
@@ -217,7 +217,7 @@ void storeTeamList(BitStream &bs, Entity *self)
         team_idx        = self->m_team->m_data.m_team_idx;
         has_taskforce   = self->m_team->m_data.m_has_taskforce;
         tm_leader_id    = self->m_team->m_data.m_team_leader_idx;
-        tm_size         = self->m_team->m_data.m_team_members.size();
+        tm_size         = (uint32_t)self->m_team->m_data.m_team_members.size();
     }
 
     storePackedBitsConditional(bs,20,team_idx);
@@ -521,7 +521,7 @@ void storePowerInfoUpdate(BitStream &bs,Entity *e)
             bs.StorePackedBits(5, power.m_level_bought);
             bs.StorePackedBits(3, power.m_charges_remaining);
             bs.StoreFloat(power.m_usage_time);
-            bs.StorePackedBits(24, power.m_activate_period);
+            bs.StorePackedBits(24, static_cast<uint32_t>(power.m_activate_period));
 
             sCDebug(logPowers) << "  NumOfEnhancements:" << power.m_total_eh_slots << "/" << power.m_enhancements.size();
             if(power.m_total_eh_slots > power.m_enhancements.size())
