+++ Servers/MapServer/NetCommandManager.cpp
@@ -140,7 +140,7 @@ void NetCommandManager::UpdateCommandShortcuts(MapClientSession *client, Vector<
     commands.reserve(m_commands_level0.size());
     for(size_t i=0, total = m_commands_level0.size(); i<total; ++i)
     {
-        client->AddShortcut(i,m_commands_level0[i]);
+        client->AddShortcut((int)i,m_commands_level0[i]);
         commands.push_back(m_commands_level0[i]->m_name);
     }
 }
